requirements.txt
psutil>=5.9.0
pyyaml>=6.0
requests>=2.28.0